+++ src/js_mixed.hpp
@@ -16,23 +16,17 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#include <common/type_deduction.hpp>
-#include <common/types.hpp>
 #include <map>
-#include <realm/mixed.hpp>
 #include <type_traits>
+
+#include "common/type_deduction.hpp"
+#include "common/types.hpp"
+#include "common/mixed_type.hpp"
 
 #pragma once
 
 namespace realm {
 namespace js {
-
-template <typename Context, typename Value>
-class MixedWrapper {
-   public:
-    virtual Mixed wrap(Context, Value const &) = 0;
-    virtual Value unwrap(Context, Mixed) = 0;
-};
 
 template <typename Context, typename Value, typename Utils>
 class MixedString : public MixedWrapper<Context, Value> {
@@ -101,6 +95,17 @@ class MixedObjectID : public MixedWrapper<Context, Value> {
 };
 
 template <typename Context, typename Value, typename Utils>
+class MixedUUID : public MixedWrapper<Context, Value> {
+    Mixed wrap(Context context, Value const &value) {
+        return Mixed(Utils::to_uuid(context, value));
+    }
+
+    Value unwrap(Context context, Mixed mixed) {
+        return Utils::from_uuid(context, mixed.get<UUID>());
+    }
+};
+
+template <typename Context, typename Value, typename Utils>
 class MixedBinary : public MixedWrapper<Context, Value> {
    private:
     // Same as with string, we need to keep this data stored on memory until the
@@ -142,7 +147,7 @@ class TypeMixed {
     using Context = typename JavascriptEngine::Context;
     using Value = typename JavascriptEngine::Value;
     using Utils = js::Value<JavascriptEngine>;
-    using Strategy = MixedWrapper<Context, Value> *;
+    using Strategy = MixedWrapper<Context, Value>;
 
     /*
         This table acts as a global hashmap,
@@ -150,7 +155,7 @@ class TypeMixed {
 
         All these pointer will be deallocated when the process exits.
     */
-    std::map<types::Type, Strategy> strategies = {
+    std::map<types::Type, Strategy *> strategies = {
         {types::String, new MixedString<Context, Value, Utils>},
         {types::Integer, new MixedNumber<Context, Value, Utils, Int>},
         {types::Float, new MixedNumber<Context, Value, Utils, Float>},
@@ -158,11 +163,10 @@ class TypeMixed {
         {types::Boolean, new MixedBoolean<Context, Value, Utils>},
         {types::Decimal, new MixedDecimal128<Context, Value, Utils>},
         {types::ObjectId, new MixedObjectID<Context, Value, Utils>},
+        {types::UUID, new MixedUUID<Context, Value, Utils>},
         {types::Binary, new MixedBinary<Context, Value, Utils>},
         {types::Timestamp, new MixedTimeStamp<Context, Value, Utils>},
     };
-
-    Strategy get_strategy(types::Type type) { return strategies[type]; }
 
     TypeMixed() {}
 
@@ -172,10 +176,18 @@ class TypeMixed {
         return instance;
     }
 
+    void unregister(types::Type type) {
+        strategies.erase(type);
+    }
+
+    void register_strategy(types::Type type, Strategy* strategy) {
+        strategies.insert( std::pair<types::Type, Strategy*>(type, strategy));
+    }
+
     Value wrap(Context context, Mixed mixed) {
         auto type_deduction = TypeDeduction::get_instance();
         auto rjs_type = type_deduction.from(mixed.get_type());
-        auto strategy = get_strategy(rjs_type);
+        auto strategy = strategies[rjs_type];
 
         if (strategy == nullptr) {
             throw std::runtime_error(
@@ -188,7 +200,7 @@ class TypeMixed {
     Mixed unwrap(Context context, Value const &js_value) {
         auto type_deduction = TypeDeduction::get_instance();
         auto type = type_deduction.typeof<JavascriptEngine>(context, js_value);
-        auto strategy = get_strategy(type);
+        auto strategy = strategies[type];
 
         if (strategy == nullptr) {
             throw std::runtime_error(