+++ tests/js/mixed-tests.js
@@ -19,16 +19,20 @@
 'use strict';
 
 const Realm = require('realm');
-let TestCase = require('./asserts');
-let {Decimal128, ObjectId} = require('bson')
+const TestCase = require('./asserts');
 const AppConfig = require('./support/testConfig');
 
+const {Decimal128, ObjectId, UUID} = Realm.BSON;
+
+
+
 const SingleSchema = {
-    name: 'Mixed',
+    name: 'mixed',
     properties: {
         a: 'mixed',
         b: 'mixed',
-        c: 'mixed'
+        c: 'mixed',
+        d: 'mixed'
     }
 }
 
@@ -43,33 +47,54 @@ module.exports = {
         TestCase.assertEqual(data.c, true, 'should store boolean (true)');
     },
 
+    testMixedQuery() {
+        const N = 100
+        let realm = new Realm({schema: [SingleSchema]})
+        for(let i=0; i < N; i++) {
+            realm.write(() => realm.create(SingleSchema.name, {a: 'xxxxxx', b: 'xxx', c: i}))
+        }
+
+        let data = realm.objects(SingleSchema.name)
+        let section = data.filtered(`a BEGINSWITH "x" AND b BEGINSWITH "x"`)
+        let half = data.filtered(`c < 50`)
+
+        TestCase.assertEqual(section.length, N, 'We expect only 100 items.');
+        TestCase.assertEqual(half.length, N/2, 'We expect only 50 items.');
+    },
+
     testMixedComplexTypes() {
         let realm = new Realm({schema: [SingleSchema]});
         let d128 = Decimal128.fromString('6.022e23');
-        let id = new ObjectId();
+        let oid = new ObjectId();
+        let uuid = new UUID();
         let date = new Date();
 
-        realm.write(()=> realm.create(SingleSchema.name, { a: id, b:d128, c: date }  ))
+        realm.write(()=> realm.create(SingleSchema.name, { a: oid, b: uuid, c: d128, d: date }  ))
 
         let data = realm.objects(SingleSchema.name)[0]
-        TestCase.assertTrue(typeof data.a === typeof id , 'should be the same type ObjectId');
-        TestCase.assertEqual(data.a.toString(), id.toString(), 'should have the same content');
-
-        TestCase.assertEqual(data.b.toString(), d128.toString(), 'Should be the same Decimal128');
-        TestCase.assertEqual(data.c.toString(), date.toString(), 'Should be the same Date');
+        TestCase.assertTrue(typeof data.a === typeof oid , 'should be the same type ObjectId');
+        TestCase.assertEqual(data.a.toString(), oid.toString(), 'should be the same ObjectId');
+        TestCase.assertEqual(data.b.toString(), uuid.toString(), 'Should be the same UUID');
+        TestCase.assertEqual(data.c.toString(), d128.toString(), 'Should be the same Decimal128');
+        TestCase.assertEqual(data.d.toString(), date.toString(), 'Should be the same Date');
     },
 
     testMixedMutability() {
         let realm = new Realm({schema: [SingleSchema]});
         let d128 = Decimal128.fromString('6.022e23');
-        let id = new ObjectId();
+        let oid = new ObjectId();
+        let uuid = new UUID();
         let date = new Date();
 
-        realm.write(()=> realm.create(SingleSchema.name, { a: id }  ))
+        realm.write(()=> realm.create(SingleSchema.name, { a: oid }))
         let data = realm.objects(SingleSchema.name)[0]
 
-        TestCase.assertTrue(typeof data.a === typeof id , 'should be the same type ObjectId');
-        TestCase.assertEqual(data.a.toString(), id.toString(), 'should have the same content');
+        TestCase.assertTrue(typeof data.a === typeof oid , 'should be the same type ObjectId');
+        TestCase.assertEqual(data.a.toString(), oid.toString(), 'should have the same content');
+
+        realm.write(()=>  data.a = uuid);
+        TestCase.assertTrue(typeof data.a === typeof uuid , 'should be the same type UUID');
+        TestCase.assertEqual(data.a.toString(), uuid.toString(), 'should have the same content UUID');
 
         realm.write(()=>  data.a = d128   )
         TestCase.assertEqual(data.a.toString(), d128.toString(), 'Should be the same Decimal128');
@@ -84,12 +109,71 @@ module.exports = {
         TestCase.assertEqual(data.a, null, 'Should be the same null');
     },
 
+    testMixedRelationalCapabilities() {
+        const VertexSchema = {
+            name: 'Vertex',
+            properties: {
+                a: 'int',
+                b: 'int',
+                c: 'int'
+            }
+        }
+
+        const MixNestedSchema = {
+            name: 'Nested',
+            properties: {
+                a: 'mixed',
+                b: 'mixed',
+                c: 'mixed'
+            }
+        }
+
+
+
+        let realm = new Realm({schema: [SingleSchema, VertexSchema, MixNestedSchema]})
+   
+        realm.write(() => {
+           let r =  realm.create(VertexSchema.name,{a:1, b: 0,  c: 0})
+           let r2 = realm.create(VertexSchema.name,{a:0, b: 1,  c: 0})
+           let r3 = realm.create(VertexSchema.name,{a:0, b: 0,  c: 1})
+
+           realm.create(SingleSchema.name, { a: r, b: r2, c: r3 } ) 
+        })
+
+        let data = realm.objects(SingleSchema.name)[0]
+        TestCase.assertEqual(data.a.a, 1, 'Should be equal 1');
+        TestCase.assertEqual(data.a.b, 0, 'Should be equal 0');
+        TestCase.assertEqual(data.a.c, 0, 'Should be equal 0');
+
+        TestCase.assertEqual(data.b.a, 0, 'Should be equal 0');
+        TestCase.assertEqual(data.b.b, 1, 'Should be equal 1');
+        TestCase.assertEqual(data.b.c, 0, 'Should be equal 0');
+
+        TestCase.assertEqual(data.c.a, 0, 'Should be equal 0');
+        TestCase.assertEqual(data.c.b, 0, 'Should be equal 0');
+        TestCase.assertEqual(data.c.c, 1, 'Should be equal 1');
+
+        realm.write(() => {
+           let r =  realm.create(MixNestedSchema.name,{a:0, b: -1})
+           let r1 = realm.create(MixNestedSchema.name,{a:1, b: 0})
+           realm.create(SingleSchema.name, { a: r, b: r1 } ) 
+        })
+
+        data = realm.objects(SingleSchema.name)[1]
+
+        TestCase.assertEqual(data.a.a,  0, 'Should be equal 0');
+        TestCase.assertEqual(data.a.b, -1, 'Should be equal -1');
+
+        TestCase.assertEqual(data.b.a, 1, 'Should be equal 1');
+        TestCase.assertEqual(data.b.b, 0, 'Should be equal 0');
+    },
+
     testMixedWrongType() {
         let realm = new Realm({schema: [SingleSchema]});
 
         TestCase.assertThrowsException(
             () => realm.write(()=> realm.create(SingleSchema.name, { a: Object.create({}) }  ) ),
-            new Error('Mixed conversion not possible for type: Object') )
+            new Error('Only Realm instances are supported.') )
     },
 
     async testMixedSync() {
@@ -111,7 +195,7 @@ module.exports = {
                 name: "MixedObject",
                 primaryKey: "_id",
                 properties: {
-                    _id: "objectId?",
+                    _id: "objectId",
                     key: "string",
                     value: "mixed"
                 }
@@ -123,13 +207,14 @@ module.exports = {
         });
 
         realm.write(() => {
-            realm.create("MixedObject", { _id: new ObjectId(), key: "one", value: 1 });
-            realm.create("MixedObject", { _id: new ObjectId(), key: "two", value: "2" });
-            realm.create("MixedObject", { _id: new ObjectId(), key: "three", value: 3.0 });
+            realm.create("MixedObject", { _id: new ObjectId(), key: "1", value: 1 });
+            realm.create("MixedObject", { _id: new ObjectId(), key: "2", value: "2" });
+            realm.create("MixedObject", { _id: new ObjectId(), key: "3", value: 3.0 });
+            realm.create("MixedObject", { _id: new ObjectId(), key: "4", value: new UUID() });
         });
 
         await realm.syncSession.uploadAllLocalChanges();
-        TestCase.assertEqual(realm.objects("MixedObject").length, 3);
+        TestCase.assertEqual(realm.objects("MixedObject").length, 4);
         realm.close();
 
         Realm.deleteFile(config);
@@ -137,11 +222,12 @@ module.exports = {
         let realm2 = await Realm.open(config);
         await realm2.syncSession.downloadAllServerChanges();
 
-        let objects = realm2.objects("MixedObject");
-        TestCase.assertEqual(objects.length, 3);
+        let objects = realm2.objects("MixedObject").sorted("key");
+        TestCase.assertEqual(objects.length, 4);
         TestCase.assertTrue(typeof objects[0].value, "number");
         TestCase.assertTrue(typeof objects[1].value, "string");
         TestCase.assertTrue(typeof objects[2].value, "number");
+        TestCase.assertTrue(objects[3].value instanceof UUID, "UUID");
 
         realm2.close();
     }